requirements.txt
fastapi
uvicorn
tensorflow
pillow
scipy
python-multipart
